start on loading posts
import a bunch of components from the previous React spike. mostly not
working yet

non-ideal: the prefetches are conducted concurrently, which means e.g.
when fetching posts for a community, i can’t guarantee that the data
for the community has already been fetched. i’ve tweaked hylo-node to
have more of the community API work with slugs instead of ID’s, but
that’s just a stopgap.
https://github.com/markdalgleish/react-fetcher/issues/2 might be the
way to go.