fix third-party login
this is how OAuth works now:

1) user clicks on a provider button in a browser
2) a popup opens directly to e.g.
http://www.hylo.com/noo/login/facebook?returnDomain=http://next.hylo.com
3) returnDomain gets stored in the session
4) when the OAuth provider redirects back to e.g.
www.hylo.com/noo/login/facebook/oauth:
4a) the cookie returned from www.hylo.com is set to work across all
subdomains of hylo.com, so the newly logged-in session can also apply to
next.hylo.com
4b) the value of returnDomain is retrieved from the session so that the
popup can use opener.sendMessage and trigger the browser to continue

as a consequence, the host for hylo-redux and hylo-node have to share a
root domain. for testing against the production hylo-node with a local
dev instance of hylo-redux, i’ve created the localdev.hylo.com CNAME,
which just points to 127.0.0.1.