use person cards for project contributors list
pass context-specific menu items for people (e.g. “Remove from
project”) from ProjectContributors as a prop on PersonCards. not sure
if this is recommended, because now we’re putting template snippets
into props, but hey, it works, and it’s a continuation of the method we
started with the Dropdown component

get moderator toggling working