write a test for UserSettings
refactor section expanding/collapsing in UserSettings to use the store,
so that we can test it with server-side rendering

n.b. @robbiecarlton: i was trying to set up a test environment for this
component to test it as if it were rendering on the client-side, so we
could test component logic that involves local state. however this
turns out to not really be possible with our setup.

you use the jsdom library to trick React into thinking it’s on
the client side, but you have to do so *globally*, by setting
global.document and window and so on, and this wrecks the parts of the
test suite that are trying to test server-side rendering, because there
are a bunch of places both in our code and in libraries that determine
whether they’re on the server or not by checking for global.window. and
we can’t just set and unset those globals before & after an individual
test, because for jsdom to affect correctly, those globals have to be
set before React is imported for the first time.

so here's another reason to err on the side of using props and actions
for components rather than local state.