fix display of comments
n.b. @robbiecarlton: the trigger for whether to fetch comments or not
is whether props.comments is null (see line 57); so if we initialize it
to an empty array the fetching never takes place. we also don’t want to
change line 57 to an isEmpty check, because then we’ll cause repeated
fetches for posts with 0 comments