move config to src/config.js
for production, we need to run babel in a build step during deploy
rather than using babel-node, which is not performant; so we need to
move all the JS that needs to be converted into the src/ directory to
use the babel —out-dir option to convert everything in one go. the only
important file for production that wasn’t already in src/ is config.