make All Posts and single-post pages into cover-image-styled pages
have the single-post page use the banner image of its first community —
this requires a reworking of the way the current community is set; it
can’t just be derived from the path since single-post pages don’t have
a community slug in their path. experiment with dispatching a
setCurrentCommunityId action in prefetch for the appropriate
components.

this is not totally satisfying, since it requires the use of
store.getState() in prefetch — it exposes as problematic the fact that
our fetch actions return two different things depending on whether
they’re cached or not