revise comment css
the strategy of making the first and last paragraphs “display: inline”,
in order to flow with the commenter’s name and the “Show more” link,
doesn’t work when there are exactly two paragraphs in the comment,
because they then flow into each other.

so instead we just move the commenter’s name into the first paragraph
and remove “display: inline” on it.

n.b. @robbiecarlton