better solution for dropdown weirdness
i originally introduced “touchend” instead of “click” on mobile because
the backdrop wasn’t sending “click” events — but that caused a race
condition between the touchend and click events on links in the
dropdown. changing the backdrop to an A tag causes it to send click
events, so we can just handle click instead of touchend and avoid the
race condition, having to filter out dragging, etc.