complete feature flags implementation
there’s some subtlety involved in getting it to work universally. we
can’t rely on envify to get all the feature flag env variables copied
over, because they’re not explicitly listed in the source code. so we
bundle them up in appHandler and pass them to the client like we do
with the initial state and asset manifest