set up a browser-based testing environment
This allows you to test client-side behavior of components in a real
browser environment (thanks to Nightmare, which wraps Electron, which
wraps Chromium).

You use this by writing a “scaffold” file under test/browser/scaffolds,
and then referencing it with the useScaffold helper method. That takes
the file and uses it as the entry point for a browserify bundle, then
creates an HTML file that loads that bundle, and runs it in Nightmare.

See test/browser/meta.test.js for example usage.